+++ tests/extra/job_management/test_job_management.py
@@ -977,6 +977,26 @@ class TestCsvJobDatabase:
         )
         assert set(db.read()["some_number"]) == {1, 2, 3}
 
+    def test_read_with_crs_column(self, tmp_path):
+        """
+        Having a column named "crs" can cause obscure error messages when creating a GeoPandas dataframe
+        https://github.com/Open-EO/openeo-python-client/issues/714
+        """
+        source_df = pd.DataFrame(
+            {
+                "crs": [1234],
+                "geometry": ["Point(2 3)"],
+            }
+        )
+        path = tmp_path / "jobs.csv"
+        source_df.to_csv(path, index=False)
+        result_df = CsvJobDatabase(path).read()
+        assert isinstance(result_df, geopandas.GeoDataFrame)
+        assert result_df.to_dict(orient="list") == {
+            "crs": [1234],
+            "geometry": [shapely.geometry.Point(2, 3)],
+        }
+
 
 class TestParquetJobDatabase: